FelixEnvironment

Documentation

data Value

Constructors for various kinds of values

Constructors

DoubleValue Double
Double value constructor

BooleanValue Bool
Bool value constructor

StringValue String
string value constructor

Closure Environment [ID] [Statement]
closure value constructor

PredefFunc ([Value] -> IO Value)
Predefined function

Undef
undef value constructor

type Environment = IORef EnvList

Mutable EnvList in IO Monad

data EnvList

constructors for envlist.

Constructors

Top (Map ID Value)
Top constructor storing map of id and value

Cons (Map ID Value) Environment
storing map of id and value alongwith mutable Envlist.

envLookup :: Environment -> ID -> IO Value

"envLookup" function look for value of id inside Environment

envDeclare :: Environment -> ID -> Value -> IO ()

"envDeclare" function insert (id,value) in mapping present in Environment.

envAssign :: Environment -> ID -> Value -> IO ()

"envAssign" function assign value to id present in environment mapping. If that Id is not present in the Environment then it stores error message in Felixerror string.

newConsEnv :: Environment -> IO Environment

"newConsEnv" creates new empty cons environment.

newConsEnvFun :: Environment -> [(ID, Value)] -> IO Environment

"newConsEnvFun" creates new cons environment and add given mapping pairs.