FelixEvaluate

Documentation

evalExpr :: Environment -> Expression -> IO Value

"evalExpr" function evaluates an expression for each type of expression defined in Syntax.hs

evalOp :: OP -> Value -> Value -> Value

"evalExpr" function evaluates an Operator expression for all the operators.

evalStmt :: Environment -> Statement -> IO (Maybe Value)

Evaluates Variable Decleration Statment. First evaluates expression and then declares this ident,value pair in environment

"evalExpr" function evaluates Statement for each type of statement defined in Syntax.hs

run :: Statements -> Environment -> IO (Maybe Value)

"run" function evaluates multiple statements