FelixParser

Documentation

oprPrecedence :: OperatorTable Parser Expression

Operator precedence table for FELIX | function calls > arithmetic operations > comparison operations | > logical operators

oprParser :: Parser Expression

oprParser parses operator | result can be one of the following : | function, operator, constant, identifier

exprParser :: Parser Expression

exprParser parses expressions | result can be one of the following: | function, operator, constant, identifier

constantParser :: Parser Constant

constantParser parses constant | result can be one of the following: | double, bool, string literal

identifierParser :: Parser ID

Parses idenrifiers . | Identifiers start with lower case letter followed by any character

statementParser :: Parser Statement

statementParser parses statenent. | Statements can be a var,if,while,return or an expression

functParser :: Parser Expression

functParser parses function | fun (args ...) {statements}

felixParser :: Parser Statements

felixParser parses statements | ignores the leading whitespaces and starts parsing the statements

parseText :: Text -> Either ParseError Statements

parseText parses text into statements. This is the function that has to be | used by other modules. This ensures the end being eof.